backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sshtunnel==0.4.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
torch==2.1.0
gymnasium==0.29.1
ta==0.11.0
stable-baselines3==2.2.1
cloudpickle>=3.1.0

